+++ CHANGES.md
@@ -1,12 +1,15 @@
 # Change Log
 
-#### next release (8.7.11)
+#### next release (8.7.12)
+
+- [The next improvement]
+
+#### 8.7.11 - 2024-12-18
 
 - Explicitly set prettier tab-width
 - Move release guide from README.md to RELEASE_GUIDE.md
 - Add `clampToGround` to `KmlCatalogItemTraits` (defaults to `true`) - this is now passed to `KmlDataSource.load`. Terria no longer clamps polygon geometries to terrain manually. All clamping logic is now handled by Cesium.
 - Add `dataSourceUri` to `KmlCatalogItemTraits` - Overrides the url to use for resolving relative links and other KML network features
-- [The next improvement]
 
 #### 8.7.10 - 2024-11-29